[next] Fixes #1484: Re-Adds the Search page component (#1621)
* Added Search component and moved search.tsx contents to SearchPage component
* Added Head, Darkmode Toggle, and Banner
